interstitial reactor cleanup in prep for keep up flow
