+++ node/src/reactor/main_reactor.rs
@@ -1,6 +1,7 @@
 //! Main reactor for nodes.
 
 mod config;
+mod control;
 mod error;
 mod event;
 mod fetchers;
@@ -33,7 +34,7 @@ use crate::{
         block_proposer::{self, BlockProposer},
         block_synchronizer::{self, BlockSynchronizer},
         block_validator::{self, BlockValidator},
-        blocks_accumulator::{BlocksAccumulator, LeapInstruction, StartingWith},
+        blocks_accumulator::{BlocksAccumulator, StartingWith, SyncInstruction},
         chain_synchronizer::{self, ChainSynchronizer},
         consensus::{self, EraReport, EraSupervisor, HighwayProtocol},
         contract_runtime::ContractRuntime,
@@ -151,383 +152,6 @@ pub(crate) struct MainReactor {
 
     // TODO: fill from: sync_leap, block execution, block_added in blocks accumulator
     validator_matrix: ValidatorMatrix,
-}
-
-enum CatchUpInstruction {
-    Do(Effects<MainEvent>),
-    CheckSoon(String),
-    CheckLater(String, u64),
-    Shutdown(String),
-    CaughtUp,
-    // TODO: These two could potentially be handled directly here
-    // Genesis,
-    // Upgrade,
-}
-
-impl MainReactor {
-    fn catch_up_instructions(
-        &mut self,
-        rng: &mut NodeRng,
-        effect_builder: EffectBuilder<MainEvent>,
-    ) -> CatchUpInstruction {
-        let mut effects = Effects::new();
-        // check idleness & enforce re-attempts if necessary
-        if let Some(timestamp) = self.block_synchronizer.last_progress() {
-            if Timestamp::now().saturating_diff(timestamp) <= self.idle_tolerances {
-                self.attempts = 0; // if any progress has been made, reset attempts
-                return CatchUpInstruction::CheckLater(
-                    "block_synchronizer is making progress".to_string(),
-                    Self::WAIT_SEC * 2,
-                );
-            }
-            self.attempts += 1;
-            if self.attempts > self.max_attempts {
-                return CatchUpInstruction::Shutdown(
-                    "catch up process exceeds idle tolerances".to_string(),
-                );
-            }
-        }
-
-        // determine which block / block_hash we should attempt to leap from
-        let starting_with = match self.trusted_hash {
-            None => {
-                match self.linear_chain.highest_block() {
-                    // no trusted hash provided use local tip if available
-                    Some(block) => {
-                        // -+ : leap w/ local tip
-                        StartingWith::Block(Box::new(block.clone()))
-                    }
-                    None => {
-                        // TODO - consider avoiding multiple commit_genesis calls.
-
-                        // if we are pre-genesis, attempt to apply genesis; if we are not shutdown
-                        match maybe_pre_genesis(
-                            effect_builder,
-                            self.chainspec.clone(),
-                            self.chainspec_raw_bytes.clone(),
-                        ) {
-                            Ok(effects) => {
-                                return CatchUpInstruction::Do(effects);
-                            }
-                            Err(msg) => {
-                                // we are post genesis, have no local blocks, and no trusted hash
-                                // so we can't possibly catch up to the network and should shut down
-                                return CatchUpInstruction::Shutdown(msg);
-                            }
-                        }
-                    }
-                }
-            }
-            Some(trusted_hash) => {
-                match self.storage.read_block(&trusted_hash) {
-                    Ok(Some(trusted_block)) => {
-                        match self.linear_chain.highest_block() {
-                            Some(block) => {
-                                // ++ : leap w/ the higher of local tip or trusted hash
-                                if trusted_block.height() > block.height() {
-                                    StartingWith::Hash(trusted_hash)
-                                } else {
-                                    StartingWith::Block(Box::new(block.clone()))
-                                }
-                            }
-                            None => {
-                                // should be unreachable if we've gotten this far
-                                StartingWith::Hash(trusted_hash)
-                            }
-                        }
-                    }
-                    Ok(None) => {
-                        // +- : leap w/ config hash
-                        StartingWith::Hash(trusted_hash)
-                    }
-                    Err(_) => {
-                        return CatchUpInstruction::Shutdown(
-                            "fatal block store error when attempting to read block under trusted hash".to_string(),
-                        );
-                    }
-                }
-            }
-        };
-
-        // the block accumulator should be receiving blocks via gossiping
-        // and usually has some awareness of the chain ahead of our tip
-        let trusted_hash = *starting_with.block_hash();
-        match self.blocks_accumulator.should_leap(starting_with) {
-            LeapInstruction::Leap => {
-                let peers_to_ask = self.small_network.peers_random_vec(
-                    rng,
-                    self.chainspec
-                        .core_config
-                        .sync_leap_simultaneous_peer_requests,
-                );
-                effects.extend(effect_builder.immediately().event(move |_| {
-                    MainEvent::SyncLeaper(sync_leaper::Event::AttemptLeap {
-                        trusted_hash,
-                        peers_to_ask,
-                    })
-                }));
-                return CatchUpInstruction::Do(effects);
-            }
-            LeapInstruction::BlockSync {
-                block_hash,
-                should_fetch_execution_state,
-            } => {
-                self.block_synchronizer.sync(
-                    block_hash,
-                    should_fetch_execution_state,
-                    self.chainspec
-                        .core_config
-                        .sync_leap_simultaneous_peer_requests,
-                );
-                return CatchUpInstruction::CheckSoon(
-                    "block_synchronizer is initialized".to_string(),
-                );
-            }
-            LeapInstruction::CaughtUp => {
-                match self.linear_chain.highest_block() {
-                    Some(block) => {
-                        let maybe_upgrade = maybe_upgrade(
-                            effect_builder,
-                            block,
-                            self.chainspec.clone(),
-                            self.chainspec_raw_bytes.clone(),
-                        );
-                        if let Ok(Some(effects)) = maybe_upgrade {
-                            return CatchUpInstruction::Do(effects);
-                        }
-                        if let Err(msg) = maybe_upgrade {
-                            return CatchUpInstruction::Shutdown(msg);
-                        }
-                    }
-                    None => {
-                        // should be unreachable
-                        return CatchUpInstruction::Shutdown(
-                            "can't be caught up with no block in the block store".to_string(),
-                        );
-                    }
-                }
-            }
-        }
-
-        // there are no catch up or shutdown instructions, so we must be caught up
-        CatchUpInstruction::CaughtUp
-    }
-
-    // put in the config
-    const WAIT_SEC: u64 = 5;
-
-    fn check_status(
-        &mut self,
-        effect_builder: EffectBuilder<MainEvent>,
-        rng: &mut NodeRng,
-    ) -> Effects<MainEvent> {
-        let effects = Effects::new();
-        match self.state {
-            ReactorState::Initialize => {
-                if let Some(effects) = initialize_component(
-                    effect_builder,
-                    &mut self.diagnostics_port,
-                    "diagnostics",
-                    MainEvent::DiagnosticsPort(diagnostics_port::Event::Initialize),
-                ) {
-                    return effects;
-                }
-                if let Some(effects) = initialize_component(
-                    effect_builder,
-                    &mut self.upgrade_watcher,
-                    "upgrade_watcher",
-                    MainEvent::UpgradeWatcher(upgrade_watcher::Event::Initialize),
-                ) {
-                    return effects;
-                }
-                if let Some(effects) = initialize_component(
-                    effect_builder,
-                    &mut self.small_network,
-                    "small_network",
-                    MainEvent::SmallNetwork(small_network::Event::Initialize),
-                ) {
-                    return effects;
-                }
-                if let Some(effects) = initialize_component(
-                    effect_builder,
-                    &mut self.event_stream_server,
-                    "event_stream_server",
-                    MainEvent::EventStreamServer(event_stream_server::Event::Initialize),
-                ) {
-                    return effects;
-                }
-                if let Some(effects) = initialize_component(
-                    effect_builder,
-                    &mut self.rest_server,
-                    "rest_server",
-                    MainEvent::RestServer(rest_server::Event::Initialize),
-                ) {
-                    return effects;
-                }
-                if let Some(effects) = initialize_component(
-                    effect_builder,
-                    &mut self.rpc_server,
-                    "rpc_server",
-                    MainEvent::RpcServer(rpc_server::Event::Initialize),
-                ) {
-                    return effects;
-                }
-                self.state = ReactorState::CatchUp;
-                return effect_builder
-                    .immediately()
-                    .event(|_| MainEvent::CheckStatus);
-            }
-            ReactorState::CatchUp => match self.catch_up_instructions(rng, effect_builder) {
-                CatchUpInstruction::Do(effects) => {
-                    let mut ret = Effects::new();
-                    ret.extend(effects);
-                    ret.extend(
-                        effect_builder
-                            .set_timeout(Duration::from_secs(Self::WAIT_SEC))
-                            .event(|_| MainEvent::CheckStatus),
-                    );
-                    return ret;
-                }
-                CatchUpInstruction::CheckLater(_, wait) => {
-                    return effect_builder
-                        .set_timeout(Duration::from_secs(wait))
-                        .event(|_| MainEvent::CheckStatus);
-                }
-                CatchUpInstruction::CheckSoon(_) => {
-                    // TODO - do we mean immediately?
-                    return effect_builder
-                        .set_timeout(Duration::from_secs(Self::WAIT_SEC))
-                        .event(|_| MainEvent::CheckStatus);
-                }
-                CatchUpInstruction::Shutdown(msg) => {
-                    return effect_builder
-                        .immediately()
-                        .event(move |_| MainEvent::Shutdown(msg))
-                }
-                CatchUpInstruction::CaughtUp => {
-                    self.state = ReactorState::KeepUp;
-                    return effect_builder
-                        .immediately()
-                        .event(|_| MainEvent::CheckStatus);
-                }
-            },
-            ReactorState::KeepUp => {
-                // TODO: define KeepUpInstruction::??? enum similar
-                // to CatchupInstruction::??? for the below steps
-                // TODO: if UpgradeWatcher announcement raised, keep track of era id's against the
-                //       new activation point
-                // detected upgrade to make this a stronger check
-                // if in validator set and era supervisor is green, validate
-                // else get added blocks for block accumulator and execute them
-                // if falling behind, switch over to catchup
-                // if sync to genesis == true, if cycles available get next historical block
-                // try real hard to stay in this mode
-                // in case of fkup, shutdown
-            }
-        }
-        effects
-
-        // TODO: Stall detection should possibly be done in the control logic.
-    }
-
-    fn handle_commit_genesis_result(
-        &mut self,
-        effect_builder: EffectBuilder<MainEvent>,
-        result: Result<GenesisSuccess, engine_state::Error>,
-    ) -> Effects<MainEvent> {
-        match result {
-            Ok(GenesisSuccess {
-                post_state_hash, ..
-            }) => {
-                info!(
-                    %post_state_hash,
-                    network_name = %self.chainspec.network_config.name,
-                    "successfully ran genesis"
-                );
-
-                let genesis_timestamp = match self
-                    .chainspec
-                    .protocol_config
-                    .activation_point
-                    .genesis_timestamp()
-                {
-                    None => {
-                        return enqueue_shutdown("must have genesis timestamp");
-                    }
-                    Some(timestamp) => timestamp,
-                };
-
-                let next_block_height = 0;
-                let initial_pre_state = ExecutionPreState::new(
-                    next_block_height,
-                    post_state_hash,
-                    BlockHash::default(),
-                    Digest::default(),
-                );
-                self.contract_runtime.set_initial_state(initial_pre_state);
-
-                let finalized_block = FinalizedBlock::new(
-                    BlockPayload::default(),
-                    Some(EraReport::default()),
-                    genesis_timestamp,
-                    EraId::default(),
-                    next_block_height,
-                    PublicKey::System,
-                );
-
-                effect_builder
-                    .enqueue_block_for_execution(finalized_block, vec![], vec![])
-                    .ignore()
-            }
-            Err(error) => enqueue_shutdown(error),
-        }
-    }
-
-    fn handle_upgrade_result(
-        &mut self,
-        effect_builder: EffectBuilder<MainEvent>,
-        previous_block_header: Box<BlockHeader>,
-        result: Result<UpgradeSuccess, engine_state::Error>,
-    ) -> Effects<MainEvent> {
-        match result {
-            Ok(UpgradeSuccess {
-                post_state_hash, ..
-            }) => {
-                info!(
-                    network_name = %self.chainspec.network_config.name,
-                    %post_state_hash,
-                    "upgrade committed"
-                );
-
-                let next_block_height = previous_block_header.height() + 1;
-                let initial_pre_state = ExecutionPreState::new(
-                    next_block_height,
-                    post_state_hash,
-                    previous_block_header.hash(),
-                    previous_block_header.accumulated_seed(),
-                );
-                self.contract_runtime.set_initial_state(initial_pre_state);
-
-                let finalized_block = FinalizedBlock::new(
-                    BlockPayload::default(),
-                    Some(EraReport::default()),
-                    previous_block_header.timestamp(),
-                    previous_block_header.next_block_era_id(),
-                    next_block_height,
-                    PublicKey::System,
-                );
-
-                effect_builder
-                    .enqueue_block_for_execution(finalized_block, vec![], vec![])
-                    .ignore()
-            }
-            Err(error) => enqueue_shutdown(error),
-        }
-    }
-}
-
-fn enqueue_shutdown<T: ToString + Send + 'static>(message: T) -> Effects<MainEvent> {
-    async {}.event(move |_| MainEvent::Shutdown(message.to_string()))
 }
 
 impl reactor::Reactor for MainReactor {
