+++ node/src/reactor/main_reactor/utils.rs
@@ -1,5 +1,5 @@
 use crate::{
-    components::{blocks_accumulator::LeapInstruction, sync_leaper, InitializedComponent},
+    components::{blocks_accumulator::SyncInstruction, sync_leaper, InitializedComponent},
     effect::{EffectBuilder, EffectExt, Effects},
     reactor::main_reactor::{MainEvent, MainReactor},
     types::{ActivationPoint, Block, BlockHash, Chainspec, ChainspecRawBytes, NodeId},
@@ -8,6 +8,10 @@ use casper_execution_engine::core::engine_state::{ChainspecRegistry, UpgradeConf
 use casper_hashing::Digest;
 use casper_types::{EraId, Key, ProtocolVersion, StoredValue, Timestamp};
 use std::{borrow::Borrow, collections::BTreeMap, sync::Arc};
+
+pub(super) fn enqueue_shutdown<T: ToString + Send + 'static>(message: T) -> Effects<MainEvent> {
+    async {}.event(move |_| MainEvent::Shutdown(message.to_string()))
+}
 
 pub(super) fn initialize_component(
     effect_builder: EffectBuilder<MainEvent>,
